requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
scipy>=1.11.0
pysr>=0.13.0
selenium>=4.15.0
pillow>=10.0.0
pytest>=7.4.0
python-dotenv>=1.0.0
requests>=2.31.0
